Duke Level-7-1: Restructure; File Access: test save file
