+++ myText.txt
@@ -1,0 +1,1 @@
+helloWold
