+++ src/main/java/Duke.java
@@ -1,9 +1,11 @@
+package src.main.java;
+
 public class Duke {
 
     public static void main(String[] args) {
 
-        Action.StartDuke();
-        Action.RunDuke();
-        Action.EndDuke();
+        DukeActionFacade.StartDuke();
+        DukeActionFacade.RunDuke();
+        DukeActionFacade.EndDuke();
     }
 }
