+++ src/main/java/TaskType.java
@@ -1,3 +1,5 @@
+package src.main.java;
+
 public enum TaskType {
     TODOS, DEADLINE, EVENT;
